+++ packages/ponder/src/graphql/buildPluralField.ts
@@ -12,21 +12,11 @@ import {
   GraphQLNonNull,
   GraphQLObjectType,
   GraphQLScalarType,
-  GraphQLSchema,
   GraphQLString,
   Kind,
 } from "graphql";
-import { Knex } from "knex";
 
-import { getEntities, getUserDefinedTypes } from "../utils/helpers";
-
-type Source = { request: unknown };
-type Context = { db: Knex<Record<string, unknown>, unknown[]> };
-
-type SingularArgs = {
-  id?: string;
-};
-type SingularResolver = GraphQLFieldResolver<Source, Context, SingularArgs>;
+import type { Context, Source } from "./types";
 
 type WhereInputArg = {
   [key: string]: number | string;
@@ -48,56 +38,6 @@ const gqlScalarStringToType: { [key: string]: GraphQLScalarType | undefined } =
     String: GraphQLString,
     Boolean: GraphQLBoolean,
   };
-
-const buildGqlSchema = (userSchema: GraphQLSchema): GraphQLSchema => {
-  const userDefinedTypes = getUserDefinedTypes(userSchema);
-
-  const entityTypes = getEntities(userSchema);
-
-  const fields: { [fieldName: string]: GraphQLFieldConfig<Source, Context> } =
-    {};
-
-  for (const entityType of entityTypes) {
-    const singularFieldName =
-      entityType.name.charAt(0).toLowerCase() + entityType.name.slice(1);
-    fields[singularFieldName] = createSingularField(entityType);
-
-    const pluralFieldName = singularFieldName + "s";
-    fields[pluralFieldName] = createPluralField(entityType, userDefinedTypes);
-  }
-
-  const queryType = new GraphQLObjectType({
-    name: "Query",
-    fields: fields,
-  });
-
-  const schema = new GraphQLSchema({ query: queryType });
-
-  return schema;
-};
-
-const createSingularField = (
-  entityType: GraphQLObjectType
-): GraphQLFieldConfig<Source, Context> => {
-  const resolver: SingularResolver = async (_, args, context) => {
-    const { db } = context;
-    const { id } = args;
-    if (!id) return null;
-
-    const query = db(entityType.name).where({ id: id });
-    const records = await query;
-
-    return records[0] || null;
-  };
-
-  return {
-    type: entityType,
-    args: {
-      id: { type: new GraphQLNonNull(GraphQLID) },
-    },
-    resolve: resolver,
-  };
-};
 
 type FilterFieldResolverConfig = {
   operator: string;
@@ -156,7 +96,7 @@ const stringSuffixToResolverConfig: {
 };
 const stringFilterSuffixes = Object.keys(stringSuffixToResolverConfig);
 
-const createPluralField = (
+const buildPluralField = (
   entityType: GraphQLObjectType,
   userDefinedTypes: {
     [key: string]: GraphQLObjectType | GraphQLEnumType | undefined;
@@ -308,4 +248,4 @@ const createPluralField = (
   };
 };
 
-export { buildGqlSchema };
+export { buildPluralField };
