+++ packages/ponder/src/graphql/index.ts
@@ -1,0 +1,1 @@
+export * from "./buildGqlSchema";
