+++ packages/ponder/src/graphql/types.ts
@@ -1,0 +1,6 @@
+import type { Knex } from "knex";
+
+type Source = { request: unknown };
+type Context = { db: Knex<Record<string, unknown>, unknown[]> };
+
+export type { Context, Source };
