+++ packages/ponder/src/indexer/buildLogWorker.ts
