+++ packages/ponder/src/indexer/index.ts
@@ -1,3 +1,3 @@
+export * from "./cache";
 export * from "./executeLogs";
 export * from "./fetchLogs";
-export * from "./logCache";
